Moved setScript to Request Class, added Execute()
Moved the setScript to the Request object.

Added the execute() method to the Client class, mostly so the workflow
makes more sense from an English syntax point of view.

Reverted the open() method back to its original logic.

Also added Docs to the classes.

Improved example of usage with setScript(), and execute with multiple
params:

/**  @var $client JonnyW\PhantomJs\Client */
$client = App::make("PhantomClient");

/**  @var $request JonnyW\PhantomJs\Message\Request */
$request = $client->getMessageFactory()->createRequest();

$request->setScript("https://raw.github.com/ariya/phantomjs/master/examples/direction.js");

/**  @var $response JonnyW\PhantomJs\Message\Response */
$response = $client->getMessageFactory()->createResponse();

// Send the request
$client->execute($request, $response, "London", "Paris");

echo $response->getCommandLine();