Change to Client->Parse() & minor fix
Change to the parse() method the Client class:
Purpose of the change is to ensure none JSON object data is handled
without interfering with the JSON object parsing.

More work needs to be done on this as currently returning command line
data is returning as along with the resource object, if both are
returned in the same request, causing the parser to return as string
including unparsed JSON. Or a JSON resource object missing command line
logs. This dependant on what is sent to log first.

Also small change to the php docs, when referring to an entire namespace
from within the same namespace, the full path needs a forward slash at
the start of the path. Otherwise one could refer to the internal path
relative to the namespace.

Minor fix to the Factory Interface:
Matched the Interface's createRequest() Method's parameter order to the
concrete class's order.