Removes the dependency on the /bin hardcoded path.
The original code assumes you will want to install binaries at /bin but probably this is not the case if you follow the defaults in composer, which will install them in /vendor/bin.

You now can have the PhantomJs in /vendor/bin if you prefer, or even at any other path, and the Engine will automatically fetch it from the proper path via the PhantomInstaller\PhantomBinary helper class.